Small refactor for nested content default serializer use.
